Moving test classes to vm package
