+++ ethereumj-core/src/test/java/org/ethereum/vm/BytecodeCompilerTest.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with the ethereumJ library. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.ethereum.test;
+package org.ethereum.vm;
 
 import org.junit.Assert;
 import org.junit.Test;
